requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
requests>=2.31.0
slowapi>=0.1.9
openai>=1.30.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
pgvector>=0.2.4
playwright